rates visual update

+++ src/Views/Book.php
@@ -24,20 +24,24 @@
         echo "<a href='http://127.0.0.6/edit-book/$id'>Edit Book </a>";
         echo "<button onclick='deleteBook($id)'>Delete book</button>";
         if (isset($rates[0])) {
+            echo "<h3><i class='fas fa-star'> Rates</i><a href='http://127.0.0.6/rate-book/$id'> + Add a rate </a></h3>";
+
             foreach ($rates as $key => $rate) {
+                echo "<div class='rate'>";
                 $rateId = $rate->getId();
                 $score = $rate->getRate();
-                echo "<h3>Rates</h3>";
                 echo "<p>$score out of 5</p>";
                 $comment = $rate->getComment();
                 if ($comment !== null) {
                     echo "<p>$comment</p>";
                 }
                 $author = $rate->getAuthor()->getUsername();
-                echo "<p> Wrote by user $author</p>";
-                echo "<a href='http://127.0.0.6/edit-rate/$rateId'>Edit Rate </a>";
-                echo "<button onclick='deleteRate($rateId)'>Delete rate</button><br>";
-                echo "<a href='http://127.0.0.6/rate-book/$id'> + Add a rate </a>";
+                echo "<p> Wrote by user <span style='color:blue'>$author</span></p>";
+                if ($_SESSION['username'] === $author) {
+                    echo "<a href='http://127.0.0.6/edit-rate/$rateId'>Edit Rate </a>";
+                    echo "<button onclick='deleteRate($rateId)'>Delete rate</button><br>";
+                }
+                echo "</div>";
             }
         } else echo "<p>No rates fort this book. <a href='http://127.0.0.6/rate-book/$id'>Be the first !</a></p>"; ?>
     </main>
